renamed CardViewProjectComponent to ProjectListComponent, moved style to css and some code refactoring
