+++ web-ng/src/app/components/project-header/project-header.component.html
@@ -22,7 +22,7 @@
       (focusOut)="updateProjectTitle($event)"
       class="inline-editor"
     ></inline-editor>
-    <button mat-stroked-button (click)="redirectMainPage()">
+    <button mat-stroked-button (click)="onProjectsButtonClick()">
       <mat-icon>list_alt</mat-icon>
       <label>Projects</label>
     </button>
