+++ web-ng/src/app/routing.module.ts
@@ -20,7 +20,7 @@ import { MainPageContainerComponent } from './components/main-page-container/mai
 import { AuthGuard } from './services/auth/auth.guard';
 import { SignInPageComponent } from './components/sign-in-page/sign-in-page.component';
 import { NavigationService } from './services/navigation/navigation.service';
-import { CardViewProjectComponent } from './components/card-view-project/card-view-project.component';
+import { ProjectListComponent } from './components/project-list/project-list.component';
 
 const routes: Routes = [
   {
@@ -35,7 +35,7 @@ const routes: Routes = [
   },
   {
     path: `${NavigationService.HOME_SEGMENT}`,
-    component: CardViewProjectComponent,
+    component: ProjectListComponent,
     canActivate: [AuthGuard],
   },
 ];
