+++ web-ng/src/app/services/auth/auth.guard.ts
@@ -56,7 +56,7 @@ export class AuthGuard implements CanActivate {
       if (!user.isAuthenticated) {
         return true;
       }
-      this.navigationService.viewProjects();
+      this.navigationService.navigateToProjectList();
       return false;
     }
     if (user.isAuthenticated) {
