+++ web-ng/src/app/shared/styles/_variables.scss
@@ -38,4 +38,5 @@ $background: map-get($theme, background);
   --drawing-tools-black: #202225;
   --drawing-tools-green: #3d7d40;
   --project-list-background: #f0f1f2;
+  --white: #ffffff;
 }
